SSAO에서 차폐 로직 개선
특정 시선 각도에서 물체 주변이 과하게 차폐되는 현상을 개선

+++ Engine/Resources/Texture.cpp
@@ -225,7 +225,7 @@ TexturePtr Texture::s_blueTex = nullptr;
 
 TexturePtr Texture::Create4x4Texture(const std::vector<uint8>& colorData)
 {
-    auto texture = Texture::Create(4, 4, GL_RGBA8, GL_RGBA, GL_UNSIGNED_BYTE);
+    auto texture = Texture::Create(4, 4, GL_RGBA16F, GL_RGBA, GL_UNSIGNED_BYTE);
 
     std::vector<uint8> fullBuffer;
     fullBuffer.reserve(4 * 4 * 4);
